[docu] Add new ROOT build option: docu
This allows to build the documentation as part of the normal build
and sets the basis for integrating it into the GitHub CI. And to
test new PullRequests so that they do not break the existing
documentation or do not add new warnings to it.

You can still build the documentation as an external project, like
is done now.

+++ documentation/doxygen/makeNotebooks.sh
@@ -62,5 +62,5 @@ EOF
 # Run rest in parallel
 xargs -L 1 -P ${nJobs:-1} ${Python3_EXECUTABLE:-python3} < $inputFile
 
-rm ${inputFile}.back
+[ -e ${inputFile}.back ] && rm ${inputFile}.back
